rc 1.0.0

See merge request personal/alinas-app/aandm_backend!3

+++ src/modules/note/note.controller.ts
@@ -58,7 +58,7 @@ export class NoteController {
     summary: 'create a note',
     description: 'creates a new note',
   })
-  async createOrUpdate(
+  async create(
     @Body() inputs: CreateNoteDto,
     @UserFromRequest() user: JWTPayload,
   ): Promise<ReS<Note>> {
@@ -96,6 +96,7 @@ export class NoteController {
           { userId: user.user.id },
           { privacyMode: In([Privacy.Public, Privacy.Protected]) },
         ],
+        order: { id: 'DESC' },
       }),
     );
   }
@@ -137,7 +138,17 @@ export class NoteController {
     summary: 'update a note',
     description: 'updates a note',
   })
-  async updateNote(@Body() dto: UpdateNoteDto): Promise<ReS<Note>> {
+  async updateNote(
+    @Body() inputs: UpdateNoteDto,
+    @UserFromRequest() user: JWTPayload,
+  ): Promise<ReS<Note>> {
+    if (!user.user.id) {
+      throw new ForbiddenException('unauthenticated');
+    }
+    const dto = {
+      ...inputs,
+      lastModifiedUserId: user.user.id,
+    };
     return ReS.FromData(await this.noteService.update(dto));
   }
 
@@ -151,7 +162,19 @@ export class NoteController {
     summary: 'delete a note',
     description: 'deletes a note',
   })
-  async deleteNote(@Param('id') id: number): Promise<ReS<Note>> {
+  async deleteNote(
+    @Param('id') id: number,
+    @UserFromRequest() user: JWTPayload,
+  ): Promise<ReS<Note>> {
+    if (!user.user.id) {
+      throw new ForbiddenException('unauthenticated');
+    }
+    const entity = await this.noteService.findOne({ where: { id } });
+    if (!entity || entity.userId !== user.user.id) {
+      throw new ForbiddenException(
+        'Model not found or does not belong to user',
+      );
+    }
     return ReS.FromData(await this.noteService.delete(id));
   }
 }
